tmpl: fix package name resolving
This change fixes package name resolving by parsing the proto files with
a small regexp for their package name, and falling back to the filename (per
proto spec) if no package statement is available.

The approach was based around the idea of streaming rather than buffering
the entire file in memory, but because *bufio.Reader does not implement the
io.Seeker interface this effectively does not work. It's still better than
before, but eventually this issue should be resolved.